fix: use explicit doubles for adaptive netplay smoothing

+++ modules/moudie-emulator/android/src/main/java/expo/modules/moudieemulator/CloudflareNetplayWebSocket.kt
@@ -66,8 +66,8 @@ class CloudflareNetplayWebSocket(
               val rtt=(SystemClock.elapsedRealtime()-sent).coerceAtLeast(0L)
               val jitter=lastRtt?.let{abs(rtt-it)}?:0L
               lastRtt=rtt
-              smoothedRtt=smoothedRtt?.let{it*0.7+rtt*0.3}?:rtt.toDouble()
-              smoothedJitter=smoothedJitter?.let{it*0.7+jitter*0.3}?:jitter.toDouble()
+              smoothedRtt=smoothedRtt?.let{it*0.7+rtt.toDouble()*0.3}?:rtt.toDouble()
+              smoothedJitter=smoothedJitter?.let{it*0.7+jitter.toDouble()*0.3}?:jitter.toDouble()
               val delay=((smoothedRtt?:rtt)+(smoothedJitter?:jitter)*1.5+16.0)/16.667
               onQuality(NetplayQuality(rttMs=(smoothedRtt?:rtt).toLong(),jitterMs=(smoothedJitter?:jitter).toLong(),probeLossPercent=0,grade="CONNECTED",recommendedDelay=delay.toLong().coerceIn(2L,45L)))
             }
